+++ src/sections/sections.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateSectionInput } from './dto/create-section.input';
@@ -12,23 +12,99 @@ export class SectionsService {
 
 
   create(createSectionInput: CreateSectionInput) {
-    const newSectionInput = this.sectionRepository.create(createSectionInput)
-    return  this.sectionRepository.save(newSectionInput)
+    try{
+      const newSectionInput = this.sectionRepository.create(createSectionInput)
+      return  this.sectionRepository.save(newSectionInput)
+    }catch(e){
+      throw new HttpException("An Error have when finding data ", HttpStatus.INTERNAL_SERVER_ERROR, { cause: e })
+    }
+   
   }
 
   findAll() {
-    return this.sectionRepository.find();
+    try{
+
+      return this.sectionRepository.find({relations:['categories' , 'candidates' , 'programmes' , 'candidates.team' , 'candidates.candidateProgrammes']});
+    }catch(e){
+      throw new HttpException("An Error have when finding data ", HttpStatus.INTERNAL_SERVER_ERROR, { cause: e })
+    }
   }
 
   findOne(id: number) {
-    return this.sectionRepository.findOneBy({id})
+    if (!id) {
+      throw new  HttpException(`section cannot be undefined`, HttpStatus.BAD_REQUEST)
+    }
+    try {
+      const section = this.sectionRepository.findOne({
+        where: {
+          id
+        },
+        relations:['categories' , 'candidates' , 'programmes' , 'candidates.team' , 'candidates.candidateProgrammes']
+      });
+      if (!section) {
+        throw new HttpException(`can't find section with id ${id}`, HttpStatus.BAD_REQUEST)
+      }
+
+      return section
+    } catch (e) {
+
+      throw new HttpException("An Error have when finding data ", HttpStatus.INTERNAL_SERVER_ERROR, { cause: e })
+
+    }
   }
 
-  update(id: number, updateSectionInput: UpdateSectionInput) {
+  findOneByName(name: string) {
+    if (!name) {
+      throw new  HttpException(`section cannot be undefined`, HttpStatus.BAD_REQUEST)
+    }
+    try {
+      const section = this.sectionRepository.findOne({
+        where: {
+          name
+        },
+        relations:['categories' , 'candidates' , 'programmes' , 'candidates.team' , 'candidates.candidateProgrammes']
+      });
+      if (!section) {
+        throw new HttpException(`can't find section with id ${name}`, HttpStatus.BAD_REQUEST)
+      }
+
+      return section
+    } catch (e) {
+
+      throw new HttpException("An Error have when finding data ", HttpStatus.INTERNAL_SERVER_ERROR, { cause: e })
+
+    }
+  }
+
+ async update(id: number, updateSectionInput: UpdateSectionInput) {
+    const section = await this.sectionRepository.findOneBy({ id })
+
+    if (!section) {
+      throw new HttpException(`Cant find a section `, HttpStatus.BAD_REQUEST)
+    }
+    // trying to return data
+
+    try {
     return this.sectionRepository.update(id, updateSectionInput);
+    } catch (e) {
+      throw new HttpException("An Error have when updating data ", HttpStatus.INTERNAL_SERVER_ERROR, { cause: e })
+    }
+
   }
 
-  remove(id: number) {
-    return this.sectionRepository.delete(id);
+ async remove(id: number) {
+    const section = await this.sectionRepository.findOneBy({ id })
+
+    if (!section) {
+      throw new HttpException(`Cant find a section `, HttpStatus.BAD_REQUEST)
+    }
+    // trying to return data
+
+    try {
+      return this.sectionRepository.delete(id);
+    } catch (e) {
+      throw new HttpException("An Error have when deleting data ", HttpStatus.INTERNAL_SERVER_ERROR, { cause: e })
+    }
+
   }
 }
