+++ src/skill/skill.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateSkillInput } from './dto/create-skill.input';
@@ -8,7 +8,7 @@ import { Skill } from './entities/skill.entity';
 @Injectable()
 export class SkillService {
 
-  constructor(@InjectRepository(Skill) private skillRepository:Repository<Skill>) {}
+  constructor(@InjectRepository(Skill) private skillRepository: Repository<Skill>) { }
 
 
   create(createSkillInput: CreateSkillInput) {
@@ -17,18 +17,87 @@ export class SkillService {
   }
 
   findAll() {
-    return this.skillRepository.find();
+    try{
+
+      return this.skillRepository.find({ relations: ['programmes'] });
+    }catch(e){
+      throw new HttpException("An Error have when finding data ", HttpStatus.INTERNAL_SERVER_ERROR, { cause: e })
+    }
+  }
+
+  findOneByName(name: string) {
+    if (!name) {
+      throw new  HttpException(`skill cannot be undefined`, HttpStatus.BAD_REQUEST)
+    }
+    try{
+      const skill = this.skillRepository.findOne({
+        where: { name },
+        relations: ['programmes','programmes.category','programmes.section']
+      })
+
+      if(!skill){
+        throw new HttpException(`Cant find skill with skill id ${name} `, HttpStatus.BAD_REQUEST)
+      }
+
+      return skill
+    }catch(e){
+      throw new HttpException("An Error have when finding data ", HttpStatus.INTERNAL_SERVER_ERROR, { cause: e })
+    }
   }
 
   findOne(id: number) {
-    return this.skillRepository.findOneBy({id})
+    if (!id) {
+      throw new  HttpException(`skill cannot be undefined`, HttpStatus.BAD_REQUEST)
+    }
+    try{
+      const skill = this.skillRepository.findOne({
+        where: { id },
+        relations: ['programmes','programmes.category','programmes.section']
+      })
+
+      if(!skill){
+        throw new HttpException(`Cant find skill with skill id ${id} `, HttpStatus.BAD_REQUEST)
+      }
+
+      return skill
+    }catch(e){
+      throw new HttpException("An Error have when finding data ", HttpStatus.INTERNAL_SERVER_ERROR, { cause: e })
+    }
+     
   }
 
   update(id: number, updateSkillInput: UpdateSkillInput) {
-    return this.skillRepository.update(id, updateSkillInput);
+    
+    const skill = this.skillRepository.findOneBy({id})
+
+    if(!skill){
+      throw new HttpException(`Cant find a skill `, HttpStatus.BAD_REQUEST)
+    }
+    // trying to return data  
+
+    try {
+      return this.skillRepository.update(id, updateSkillInput);
+    } catch (e) {
+      throw new HttpException("An Error have when updating data ", HttpStatus.INTERNAL_SERVER_ERROR, { cause: e })
+    }
+
+   
   }
 
   remove(id: number) {
+
+    const skill = this.skillRepository.findOneBy({id})
+
+    if(!skill){
+      throw new HttpException(`Cant find a skill `, HttpStatus.BAD_REQUEST)
+    }
+    // trying to return data
+
+    try {
     return this.skillRepository.delete(id);
+    }
+    catch (e) {
+      throw new HttpException("An Error have when deleting data ", HttpStatus.INTERNAL_SERVER_ERROR, { cause: e })
+    }
   }
 }
